Split BlazeClientSuite
